+++ Assets/Scenes/SampleScene.unity
@@ -374,3 +374,5 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   riverPrefab: {fileID: 8968545609573519753, guid: b8d6c4dfd008b0343b9181e181ec1741,
     type: 3}
+  roadPrefab: {fileID: 7849258118220870130, guid: 58cfbae015cb6424e91ff7d478c1f517,
+    type: 3}
